refactor(filid): split core/rules + hooks + mcp server/review-manage + module-main-analyzer
Batch A: 8 files with 4-7 functions each split into single-fn files,
originals converted to facades to preserve external import paths.

- core/rules/drift-detector (6 fn) → 4 files (fractal root)
- core/rules/document-validator (4 fn) → 3 files
- core/rules/fractal-validator (4 fn) → 3 files
- hooks/shared (4 fn) → 4 files
- mcp/tools/review-manage/utils/review-utils (5 fn) → 5 files
- mcp/tools/review-manage/format/review-format (7 fn) → 5 files
- core/module/module-main-analyzer (5 fn) → 4 files
- mcp/server/server (6 fn) → 3 files
