+++ packages/filid/src/mcp/server/start-server.ts
@@ -1,0 +1,10 @@
+import { StdioServerTransport } from '@modelcontextprotocol/sdk/server/stdio.js';
+
+import { createServer } from './create-server.js';
+
+/** Start the MCP server with stdio transport */
+export async function startServer(): Promise<void> {
+  const server = createServer();
+  const transport = new StdioServerTransport();
+  await server.connect(transport);
+}
