+++ packages/filid/src/mcp/tools/review-manage/format/collapsible.ts
@@ -1,0 +1,3 @@
+export function collapsible(summary: string, content: string): string {
+  return `<details><summary>${summary}</summary>\n\n${content}\n\n</details>`;
+}
