+++ packages/filid/src/mcp/tools/review-manage/format/format-human-summary.ts
@@ -1,0 +1,5 @@
+import type { HumanSummary } from '../../../../types/summary.js';
+
+export function formatHumanSummary(summary: HumanSummary): string {
+  return summary.markdown;
+}
