+++ packages/filid/src/mcp/tools/review-manage/utils/extract-verdict.ts
@@ -1,0 +1,6 @@
+export function extractVerdict(content: string): string {
+  const match = content.match(
+    /\*\*Verdict\*\*:\s*(APPROVED|REQUEST_CHANGES|INCONCLUSIVE)/,
+  );
+  return match?.[1] ?? 'UNKNOWN';
+}
