+++ packages/filid/src/mcp/tools/review-manage/utils/resolve-review-dir.ts
@@ -1,0 +1,14 @@
+import path from 'node:path';
+import { normalizeBranch } from './normalize-branch.js';
+
+export function resolveReviewDir(
+  projectRoot: string,
+  branchName: string,
+): string {
+  return path.join(
+    projectRoot,
+    '.filid',
+    'review',
+    normalizeBranch(branchName),
+  );
+}
