+++ packages/filid/src/mcp/tools/review-manage/utils/try-read-file.ts
@@ -1,0 +1,9 @@
+import fs from 'node:fs/promises';
+
+export async function tryReadFile(filePath: string): Promise<string | null> {
+  try {
+    return await fs.readFile(filePath, 'utf-8');
+  } catch {
+    return null;
+  }
+}
